Close template trace resources
